VoiceChannelCloser fix

+++ DiscordBot/src/main/java/com/epam/sturbun/util/VoiceChannelCloser.java
@@ -2,7 +2,6 @@ package com.epam.sturbun.util;
 
 import net.dv8tion.jda.api.entities.VoiceChannel;
 
-import java.util.concurrent.Callable;
 import java.util.concurrent.TimeUnit;
 
 public class VoiceChannelCloser implements Runnable {
@@ -15,20 +14,15 @@ public class VoiceChannelCloser implements Runnable {
 
     @Override
     public void run() {
-        Boolean isClosed = false;
-        while (!isClosed) {
-            Callable<Boolean> task = () -> {
-                TimeUnit.MINUTES.sleep(1);
-                if (toClose.getMembers().size() == 0) {
-                    toClose.delete().submit();
-                    return true;
-                } else return false;
-            };
+        Runnable task = () -> {
             try {
-                isClosed = task.call();
-            } catch (Exception e) {
+                TimeUnit.MINUTES.sleep(1);
+            } catch (InterruptedException e) {
                 e.printStackTrace();
             }
-        }
+            if (toClose.getMembers().size() == 0) toClose.delete().submit();
+        };
+        Thread closeTask = new Thread(task);
+        closeTask.start();
     }
 }
